+++ cmd/vault/config_cli.go
@@ -17,7 +17,7 @@ const (
 	tokenKeyFile     = "vault-token.key"
 	sharedTokenVault = "shared-token-vault.json"
 	saltSize         = 16
-	vaultVersion     = "0.3.3"
+	vaultVersion     = "0.3.4"
 )
 
 func showConfig() {